requirements.txt
Flask==2.2.5
Flask_Bcrypt==1.0.1
Flask_Login==0.6.2
Flask_Mail==0.9.1
flask_sqlalchemy==3.0.5
Flask_WTF==1.1.1
itsdangerous==2.1.2
SQLAlchemy==2.0.17
WTForms==3.0.1
gunicorn==20.1.0
website==1.1
python-dotenv==0.21.1
email_validator==1.3.1